Move commands & presenter files into new foldes
