+++ MvvmCross/Commands/MvxCommandCollectionBuilder.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-namespace MvvmCross.ViewModels
+namespace MvvmCross.Commands
 {
     public class MvxCommandCollectionBuilder
         : IMvxCommandCollectionBuilder
