+++ MvvmCross/Presenters/IMvxOverridePresentationAttribute.cs
@@ -2,10 +2,12 @@
 // The .NET Foundation licenses this file to you under the MS-PL license.
 // See the LICENSE file in the project root for more information.
 
-namespace MvvmCross.ViewModels
+using MvvmCross.ViewModels;
+
+namespace MvvmCross.Presenters
 {
-    public interface IMvxCommandCollectionBuilder
+    public interface IMvxOverridePresentationAttribute
     {
-        IMvxCommandCollection BuildCollectionFor(object owner);
+        MvxBasePresentationAttribute PresentationAttribute(MvxViewModelRequest request);
     }
 }
