+++ MvvmCross/Presenters/MvxAttributeViewPresenter.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MS-PL license.
 // See the LICENSE file in the project root for more information.
 
@@ -7,8 +7,9 @@ using System.Collections.Generic;
 using System.Linq;
 using MvvmCross.Logging;
 using MvvmCross.ViewModels;
+using MvvmCross.Views;
 
-namespace MvvmCross.Views
+namespace MvvmCross.Presenters
 {
     public abstract class MvxAttributeViewPresenter : MvxViewPresenter, IMvxAttributeViewPresenter
     {
