+++ MvvmCross/Presenters/MvxPresentationAttributeExtensions.cs
@@ -6,7 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace MvvmCross.Views
+namespace MvvmCross.Presenters
 {
     public static class MvxPresentationAttributeExtensions
     {
