+++ MvvmCross/ViewModels/Hints/MvxClosePresentationHint.cs
@@ -1,8 +1,8 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MS-PL license.
 // See the LICENSE file in the project root for more information.
 
-namespace MvvmCross.ViewModels
+namespace MvvmCross.ViewModels.Hints
 {
     public class MvxClosePresentationHint
         : MvxPresentationHint
